interpolation correction

+++ 0x1E-search_algorithms/102-interpolation.c
@@ -10,7 +10,7 @@
 
 int interpolation_search(int *array, size_t size, int value)
 {
-	size_t pos;
+	int pos;
 	size_t low = 0;
 	size_t high = size - 1;
 
@@ -21,7 +21,7 @@ int interpolation_search(int *array, size_t size, int value)
 	{
 		pos = low + (((double)(high - low) / (array[high] - array[low]))
 				* (value - array[low]));
-		printf("Value checked array[%lu] = [%d]\n", pos, array[pos]);
+		printf("Value checked array[%d] = [%d]\n", pos, array[pos]);
 		if (array[pos] == value)
 			return (pos);
